Categorize the dominant relationship from an ancestry match matrix

This function takes an ancestry match matrix (AMM) and it returns what the "dominant" relationship is. It doesn't try to capture all the nuances that might be present in the AMM. For example, it does not note inbreeding and other such things. However this is useful for categorizing what type of relationship is the "closest" relationship type amongst all the ways a pair is related. Currently for num_generations == 2, the function goes through in this order, to identify relationships:

Se: self. This is as far as it goes with num_generations = 0

PO: parent-offspring

Si: sibling. This is as far as it goes with num_generations = 1

GP: grandparental

A : avuncular (aunt-neice)

FC: first cousin.

cat_dom_relat(AMM)

Arguments

AMM
The ancestry match matrix to categorize

Value

This returns a list with two components:

type: a string saying what type of relationship (i.e., "Si", or "A").

hits: a two vector of the number of TRUEs in each "relationship zone" on the upper or the lower diagonal of the AMM.

Details

And so forth. This has been implemented out to 3 generations using the relationship zones in package data object relationship_zone_names.

Examples

cat_dom_relat(half_first_cousin_amm)

#> $type
#> [1] "FC"
#> 
#> $hits
#> [1] 1 1
#>