compile pairwise relationships from the samples

Run this on some of the output from slurp_spip().

compile_related_pairs(S)

Arguments

S
a tibble that must have the columns:

ID: the id of the sample

ancestors: a list column of the ancestor vectors of each individual

relatives: a list column of the vectors of individual samples (including self) that each indvidual is related to.

Value

a tibble with columns id_1 and id_2 for each pair. Any additional columns outside of relatives will be joined with _1 and _2 suffixes. There are other columns in the return too.... More later.