downsample the number of indivduals sampled

This discards individuals from the sample, randomly, until the desired number of samples is achieved, then it returns only those pairs in which both members are part of the retained samples.

downsample_pairs(S, P, n)

Arguments

S
the tibble of samples with columns at least of ID and samp_years_list. Typically this will be what is returned in the samples component from slurp_spip().

P
the tibble of pairs. Typically this will be what has been returned from compile_related_pairs().

n
The desired number of individuals (or instances, really, see below) to retain in the sample.

Value

A tibble like P except having removed any pairs that include individuals that were not retained in the sample.

Examples

# prepare some input
S <- three_pops_with_mig_slurped_results$samples
P <- compile_related_pairs(three_pops_with_mig_slurped_results$samples)
result <- downsample_pairs(S, P, n = 500)

# print the result
result

#> $ds_samples
#> # A tibble: 500 x 13
#>    ID        samp_years_list sex   born_year born_pop pop_pre pop_post pop_dur
#>    <chr>     <list>          <chr>     <int>    <int> <chr>   <chr>    <chr>  
#>  1 F10_0_28  <int [1]>       F            10        0 NA      0        NA     
#>  2 F10_0_37  <int [1]>       F            10        0 0       NA       NA     
#>  3 F10_0_50  <int [1]>       F            10        0 0       NA       NA     
#>  4 F10_0_51  <int [1]>       F            10        0 NA      0        NA     
#>  5 F10_0_60  <int [1]>       F            10        0 0       NA       NA     
#>  6 F10_0_83  <int [1]>       F            10        0 0       NA       NA     
#>  7 F10_0_86  <int [1]>       F            10        0 0       NA       NA     
#>  8 F10_1_108 <int [1]>       F            10        1 1       NA       NA     
#>  9 F10_1_12  <int [1]>       F            10        1 1       NA       NA     
#> 10 F10_1_34  <int [1]>       F            10        1 NA      1        NA     
#> # … with 490 more rows, and 5 more variables: samp_years_list_pre <list>,
#> #   samp_years_list_dur <list>, samp_years_list_post <list>, ancestors <list>,
#> #   relatives <list>
#> 
#> $ds_pairs
#> # A tibble: 183 x 33
#>    id_1  id_2  conn_comp dom_relat max_hit dr_hits upper_member times_encounter…
#>    <chr> <chr>     <dbl> <chr>       <int> <list>         <int>            <int>
#>  1 F10_… F10_…         1 Si              1 <int […           NA                2
#>  2 F10_… M10_…         1 Si              1 <int […           NA                3
#>  3 F10_… F10_…         2 Si              1 <int […           NA                2
#>  4 F10_… M10_…         2 Si              1 <int […           NA                3
#>  5 F10_… M10_…         3 Si              1 <int […           NA                2
#>  6 F10_… F10_…         4 Si              1 <int […           NA                2
#>  7 F10_… M11_…         5 Si              1 <int […           NA                3
#>  8 F10_… M7_1…         5 PO              1 <int […            2                3
#>  9 F10_… M10_…         6 Si              1 <int […           NA                3
#> 10 F10_… M10_…         6 Si              1 <int […           NA                3
#> # … with 173 more rows, and 25 more variables: primary_shared_ancestors <list>,
#> #   psa_tibs <list>, pop_pre_1 <chr>, pop_post_1 <chr>, pop_dur_1 <chr>,
#> #   pop_pre_2 <chr>, pop_post_2 <chr>, pop_dur_2 <chr>, sex_1 <chr>,
#> #   sex_2 <chr>, born_year_1 <int>, born_year_2 <int>,
#> #   samp_years_list_pre_1 <list>, samp_years_list_1 <list>,
#> #   samp_years_list_dur_1 <list>, samp_years_list_post_1 <list>,
#> #   samp_years_list_pre_2 <list>, samp_years_list_2 <list>,
#> #   samp_years_list_dur_2 <list>, samp_years_list_post_2 <list>,
#> #   born_pop_1 <int>, ancestors_1 <list>, born_pop_2 <int>, ancestors_2 <list>,
#> #   anc_match_matrix <list>
#>