compute the probability that a pair are half siblings

Just a quick function for some developmental work. This is not meant for prime time (and, in fact, most of the inference related stuff like this is going to be in a differnet package). Depends on population sizes and covariates. This is not really vectorized, but if you pass it multiple N values, it will return a vector, one for each N value. Note that we only expect this to work well for individuals with different birth years. When they have the same birth year they could have been full siblings and also variance in reproductive success comes more into play. This assumes an equal sex ratio.

half_sib_kin_probs(y, a1, a2, N, surv_probs, prob_repro, rel_repro)

Arguments

y
the difference in birth years

a1
the age at sampling of one member of the pair

a2
the age at sampling of another member of the pair

N
total number of 2+ individuals in the population. This can be a vector.

surv_probs
a vector giving the age specific survival probabilities.

prob_repro
The age-specific (starting from 1) probability of reproducing in a year

rel_repro
The relative fecundities/reproductive potential of the individuals