plot an ancestry matrix (or multiple such matrices) from its (their) matrix form

For illustration purposes, if you want to simply plot an ancestry matrix (or several) to show particular values, then this is the handy function for you.

plot_amm_from_matrix(X)

Arguments

X

input tibble with a factor or character column ID that gives the "name" of the ancestry matrix that will be used if you want to facet over the values in ID. And also X must have a list column anc_match_matrix each element of which is a logical ancestry match matrix. X may have a list column of tibbles called psa_tibs that says which cells are the primary shared ancestors.

Value

plot_amm_from_matrix() returns a ggplot object: each facet is an image of the ancestry match matrix. It is facet-wrapped over the values in the ID column of X.

Examples

# get some input: all the 2-generation AMMs in `example_amms`
X <- example_amms[stringr::str_detect(names(example_amms), "2gen$")] %>%
    tibble::enframe(name = "ID", value = "anc_match_matrix")

# plot those
g <- plot_amm_from_matrix(X) +
    ggplot2::facet_wrap(~ ID)