Return the perimeters of all the relationship zones

This is primarily for plotting a figure in the paper about this package, showing where all the relationship zones are. It merely cycles over the possible relationships in relationship_zone_names and produces one or two rows in a tibble for each that has the corners of the rectangle of that zone in the columns xmin, xmax, ymin, and ymax. It is designed to be overlaid upon the ancestry_match_matrix plots. There are some additional columns that give us the midpoint of the area, etc.

relationship_zone_perimeters()

Value

Returns a tibble with the following columns:

which_matrix: a column of values M1 or M2. M1 denotes that the row's values are for the relationship zone found in or below the lower diagonal of the ancestry match matrix and M2 denotes that the row's value are of the zone found in the upper part of the ancestry match matrix. Symmetrical relationships are considered to be M1.

zone: The abbreviation for the relationship (e.g., Se, PO, Si, etc.)

xmin: The left-hand x value of the zone.

xmax: The right-hand x value of the zone.

ymin: The bottom y value of the zone.

ymax: The top y value of the zone.

area: The area in unit squares of the zone.

xmid: The x midpoint of the zone.

ymid: The y midpoint of the zone.

Examples

relationship_zone_perimeters()

#> # A tibble: 25 x 9
#>    which_matrix zone   xmin  xmax  ymin  ymax  area  xmid  ymid
#>    <chr>        <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 M1           Se      0.5   1.5   0.5   1.5     1   1     1  
#>  2 M1           PO      1.5   3.5   0.5   1.5     2   2.5   1  
#>  3 M2           PO      0.5   1.5   1.5   3.5     2   1     2.5
#>  4 M1           Si      1.5   3.5   1.5   3.5     4   2.5   2.5
#>  5 M1           GP      3.5   7.5   0.5   1.5     4   5.5   1  
#>  6 M2           GP      0.5   1.5   3.5   7.5     4   1     5.5
#>  7 M1           A       3.5   7.5   1.5   3.5     8   5.5   2.5
#>  8 M2           A       1.5   3.5   3.5   7.5     8   2.5   5.5
#>  9 M1           FC      3.5   7.5   3.5   7.5    16   5.5   5.5
#> 10 M1           GGP     7.5  15.5   0.5   1.5     8  11.5   1  
#> # … with 15 more rows