Return a Leslie-like matrix from the spip parameters

Here we take the survival rates for females and males and the sex ratio, as well as the annual new cohort size (assumed constant), and we make a leslie-like matrix to compute the stable age distribution.

leslie_from_spip(P, C)

Arguments

P
a named list of the spip parameters.

C
the constant size of the newborn cohort each year

Value

This function returns a list with the following components:

stable_age_distro_fem: a vector of the expected number of females in each age group of 0 up to MaxAge-1 once the stable age distribution has been reached. Note that this corresponds to the PREKILL_CENSUS from spip. In this vector, the size of the MaxAge group is left out because this is how it is needed to be to insert into the --initial-males and --initial-females options in spip(). If you want the size of all age classes, use the output list component stable_age_distro_fem_with_max_age_class, described below.

stable_age_distro_male: same as above, but for males.

stable_age_distro_fem_with_max_age_class: The expected number of females from age 0 to MaxAge once the stable age distribution has been reached.

stable_age_distro_male_with_max_age_class: same as above, but for females.

female_leslie_matrix: The Leslie matrix implied by the spip parameters in P.

Examples

result <- leslie_from_spip(species_1_life_history, 300)

# print the result list:
result

#> $stable_age_distro_fem
#>  [1] 150.000000 112.500000  85.500000  64.980000  50.034600  38.526642
#>  [7]  30.050781  23.439609  18.517291  14.628660  11.702928   9.362342
#> [13]   7.489874   6.066798   4.914106   4.029567   3.304245   2.709481
#> [19]   2.221774   1.821855
#> 
#> $stable_age_distro_male
#>  [1] 150.000000 112.500000  85.500000  64.980000  50.034600  38.526642
#>  [7]  30.050781  23.439609  18.517291  14.628660  11.702928   9.362342
#> [13]   7.489874   6.066798   4.914106   4.029567   3.304245   2.709481
#> [19]   2.221774   1.821855
#> 
#> $stable_age_distro_fem_with_max_age_class
#>  [1] 150.000000 112.500000  85.500000  64.980000  50.034600  38.526642
#>  [7]  30.050781  23.439609  18.517291  14.628660  11.702928   9.362342
#> [13]   7.489874   6.066798   4.914106   4.029567   3.304245   2.709481
#> [19]   2.221774   1.821855   1.493921
#> 
#> $stable_age_distro_male_with_max_age_class
#>  [1] 150.000000 112.500000  85.500000  64.980000  50.034600  38.526642
#>  [7]  30.050781  23.439609  18.517291  14.628660  11.702928   9.362342
#> [13]   7.489874   6.066798   4.914106   4.029567   3.304245   2.709481
#> [19]   2.221774   1.821855   1.493921
#> 
#> $female_leslie_matrix
#>       [,1] [,2] [,3]       [,4]      [,5]      [,6]      [,7]      [,8]
#>  [1,] 0.00 0.00 0.00 0.01517058 0.0713358 0.3014811 0.6539862 0.8630492
#>  [2,] 0.75 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.00 0.76 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [4,] 0.00 0.00 0.76 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [5,] 0.00 0.00 0.00 0.77000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [6,] 0.00 0.00 0.00 0.00000000 0.7700000 0.0000000 0.0000000 0.0000000
#>  [7,] 0.00 0.00 0.00 0.00000000 0.0000000 0.7800000 0.0000000 0.0000000
#>  [8,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.7800000 0.0000000
#>  [9,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.7900000
#> [10,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [11,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [12,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [13,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [14,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [15,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [16,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [17,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [18,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [19,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [20,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [21,] 0.00 0.00 0.00 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           [,9]    [,10]    [,11]    [,12]    [,13]    [,14]    [,15]    [,16]
#>  [1,] 0.943583 1.000598 1.036018 1.071437 1.120692 1.156554 1.207138 1.243443
#>  [2,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [3,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [4,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [5,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [6,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [7,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [8,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [10,] 0.790000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [11,] 0.000000 0.800000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [12,] 0.000000 0.000000 0.800000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [13,] 0.000000 0.000000 0.000000 0.800000 0.000000 0.000000 0.000000 0.000000
#> [14,] 0.000000 0.000000 0.000000 0.000000 0.810000 0.000000 0.000000 0.000000
#> [15,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.810000 0.000000 0.000000
#> [16,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.820000 0.000000
#> [17,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.820000
#> [18,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [19,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [20,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [21,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>          [,17]    [,18]    [,19]    [,20] [,21]
#>  [1,] 1.279747 1.316052 1.352357 1.388662     0
#>  [2,] 0.000000 0.000000 0.000000 0.000000     0
#>  [3,] 0.000000 0.000000 0.000000 0.000000     0
#>  [4,] 0.000000 0.000000 0.000000 0.000000     0
#>  [5,] 0.000000 0.000000 0.000000 0.000000     0
#>  [6,] 0.000000 0.000000 0.000000 0.000000     0
#>  [7,] 0.000000 0.000000 0.000000 0.000000     0
#>  [8,] 0.000000 0.000000 0.000000 0.000000     0
#>  [9,] 0.000000 0.000000 0.000000 0.000000     0
#> [10,] 0.000000 0.000000 0.000000 0.000000     0
#> [11,] 0.000000 0.000000 0.000000 0.000000     0
#> [12,] 0.000000 0.000000 0.000000 0.000000     0
#> [13,] 0.000000 0.000000 0.000000 0.000000     0
#> [14,] 0.000000 0.000000 0.000000 0.000000     0
#> [15,] 0.000000 0.000000 0.000000 0.000000     0
#> [16,] 0.000000 0.000000 0.000000 0.000000     0
#> [17,] 0.000000 0.000000 0.000000 0.000000     0
#> [18,] 0.820000 0.000000 0.000000 0.000000     0
#> [19,] 0.000000 0.820000 0.000000 0.000000     0
#> [20,] 0.000000 0.000000 0.820000 0.000000     0
#> [21,] 0.000000 0.000000 0.000000 0.820000     0
#>